refactor: renames numberUtils into number
